+++ components/create-campaign/CampaignOutlook.tsx
@@ -2,39 +2,20 @@ import {
 	DummyAvatar,
 	MoneyIcon,
 	ShareIcon,
-	whatsappIcon,
-	xIcon,
 } from "@/components/common/campaign-icons";
-import type { ApiResponse } from "@/interfaces";
 import type { Campaign } from "@/interfaces/Campaign";
-import {
-	type DonationDetailsType,
-	callApi,
-	cn,
-	getDaysLeft,
-	zodValidator,
-} from "@/lib";
+import { cn, getDaysLeft } from "@/lib";
 import { getDateFromString } from "@/lib/helpers/campaign";
-import { useElementList, useShareCampaign } from "@/lib/hooks";
+import { useElementList } from "@/lib/hooks";
 import { useSlot } from "@/lib/hooks/useSlot";
-import { useSession } from "@/store";
-import { zodResolver } from "@hookform/resolvers/zod";
 import { format } from "date-fns";
-import { FilesIcon, LinkIcon } from "lucide-react";
 import Image from "next/image";
-import Link from "next/link";
-import { useEffect, useState } from "react";
-import { type SubmitHandler, useForm } from "react-hook-form";
-import { toast } from "sonner";
-import {
-	CustomDialog,
-	FormErrorMessage,
-	Heading,
-	SingleCampaignProgress,
-} from "../common";
+import { useState } from "react";
+import { Heading, SingleCampaignProgress } from "../common";
 import { FAQ, UrgentFundraisers } from "../common/landingPage";
-import { Button, Checkbox, Input } from "../ui";
+import { Button } from "../ui";
 import CampaignCarousel from "./CampaignCarousel";
+import DonationFlowDialog from "./DonationFlowDialog";
 import DonorSection from "./DonorSection";
 import ShareCampaignDialog from "./ShareCampaignDialog";
 
@@ -56,6 +37,8 @@ function CampaignOutlook(props: CampaignOutlookProps) {
 
 	const [TagList] = useElementList();
 
+	const [hideExcessStory, setHideExcessStory] = useState(true);
+
 	const HeaderSlot = useSlot(children, CampaignOutlook.Header);
 
 	const fundraiserTarget =
@@ -63,57 +46,6 @@ function CampaignOutlook(props: CampaignOutlookProps) {
 		"Beneficiary";
 
 	const campaignDeadline = getDateFromString(campaign.deadline);
-
-	//donation flow
-	const { user } = useSession((state) => state);
-	useEffect(() => {
-		if (user) {
-			reset({
-				donorName: `${user.firstName} ${user.lastName}`,
-				donorEmail: user.email,
-				amount: "",
-			});
-		}
-	}, [user]);
-	const [hideMyDetails, setHideMyDetails] = useState(false);
-	const {
-		register,
-		handleSubmit,
-		reset,
-		formState: { errors },
-	} = useForm<DonationDetailsType>({
-		resolver: zodResolver(zodValidator("donationDetails")!),
-		mode: "onChange",
-		reValidateMode: "onChange",
-		defaultValues: {
-			donorName: "",
-			donorEmail: "",
-			amount: "",
-		},
-	});
-	const onDonateSubmit: SubmitHandler<DonationDetailsType> = async (
-		data: DonationDetailsType
-	) => {
-		console.log({ ...data, hideMyDetails, campaignId, amount: +data.amount });
-		const { data: dataInfo, error } = await callApi<ApiResponse>(
-			"/donation/create",
-			{
-				...data,
-				hideMyDetails,
-				campaignId,
-				amount: +data.amount,
-			}
-		);
-		if (error) {
-			toast.error("Error", {
-				description: error.message,
-			});
-			return;
-		}
-		toast.success("Success", {
-			description: dataInfo?.message,
-		});
-	};
 
 	return (
 		<main className="flex flex-col pb-20 text-abeg-text">
@@ -146,122 +78,14 @@ function CampaignOutlook(props: CampaignOutlookProps) {
 						<figcaption className="text-lg">{fundraiserTarget}</figcaption>
 					</figure>
 
-					<CustomDialog
-						classNames={{
-							content: "gap-0 p-12 md:p-12 w-full max-w-[500px]",
-						}}
+					<DonationFlowDialog
+						campaignId={campaignId}
 						trigger={
 							<button className="rounded-[20px] border border-white bg-white/30 px-[30px] py-1 text-lg font-bold">
 								Donate
 							</button>
 						}
-					>
-						<form
-							className="flex flex-col gap-4"
-							onSubmit={(e) => {
-								e.preventDefault();
-								handleSubmit(onDonateSubmit)(e);
-							}}
-						>
-							<p className="mt-2 text-center">
-								Donate to this campaign. Every penny brings us one step closer
-								to making a difference
-							</p>
-							<div className="space-y-1 bg-white">
-								<label
-									htmlFor="donorName"
-									className="text-sm font-medium md:text-base"
-								>
-									Full name
-								</label>
-								<Input
-									{...register("donorName")}
-									autoFocus
-									type="text"
-									id="donorName"
-									required
-									placeholder="Enter your Full name"
-									className={`min-h-[45px] border text-sm font-light disabled:cursor-not-allowed disabled:bg-inherit md:text-base ${
-										errors.donorName &&
-										"ring-2 ring-abeg-error-20 placeholder:text-abeg-error-20"
-									}`}
-									disabled={user?.firstName && user.lastName ? true : false}
-								/>
-								<FormErrorMessage
-									error={errors.donorName!}
-									errorMsg={errors.donorName?.message!}
-								/>
-							</div>
-							<div className="space-y-1">
-								<label
-									htmlFor="donorEmail"
-									className="text-sm font-medium md:text-base"
-								>
-									Email address
-								</label>
-								<Input
-									{...register("donorEmail")}
-									autoFocus
-									type="text"
-									id="donorEmail"
-									required
-									placeholder="Enter your email address"
-									className={`min-h-[45px] border text-sm font-light disabled:cursor-not-allowed disabled:bg-inherit md:text-base ${
-										errors.donorEmail &&
-										"ring-2 ring-abeg-error-20 placeholder:text-abeg-error-20"
-									}`}
-									disabled={user?.email ? true : false}
-								/>
-								<FormErrorMessage
-									error={errors.donorEmail!}
-									errorMsg={errors.donorEmail?.message!}
-								/>
-							</div>
-							<div className="space-y-1">
-								<label
-									htmlFor="amount"
-									className="text-sm font-medium md:text-base"
-								>
-									Amount
-								</label>
-								<Input
-									{...register("amount")}
-									autoFocus
-									type="number"
-									id="amount"
-									required
-									placeholder="Enter any amount"
-									className={`min-h-[45px] border text-sm font-light disabled:cursor-not-allowed disabled:bg-inherit md:text-base ${
-										errors.amount &&
-										"ring-2 ring-abeg-error-20 placeholder:text-abeg-error-20"
-									}`}
-								/>
-								<FormErrorMessage
-									error={errors.amount!}
-									errorMsg={errors.amount?.message!}
-								/>
-							</div>
-							<div className="flex items-center gap-3">
-								<Checkbox
-									name="hideMyDetails"
-									id="hideMydetails"
-									className="accent-red-500"
-									checked={hideMyDetails}
-									onCheckedChange={() => setHideMyDetails(!hideMyDetails)}
-								/>
-								<label
-									htmlFor="hideMyDetails"
-									className="cursor-pointer text-sm font-medium md:text-base"
-									onClick={() => setHideMyDetails(!hideMyDetails)}
-								>
-									Donate anonymously
-								</label>
-							</div>
-							<Button className="bg-abeg-primary text-base text-white">
-								Donate
-							</Button>
-						</form>
-					</CustomDialog>
+					/>
 
 					<ShareCampaignDialog
 						campaign={campaign}
@@ -319,10 +143,8 @@ function CampaignOutlook(props: CampaignOutlookProps) {
 					/>
 
 					<article className="space-y-4">
-						<CustomDialog
-							classNames={{
-								content: "gap-0 p-12 md:p-12 w-full max-w-[500px]",
-							}}
+						<DonationFlowDialog
+							campaignId={campaignId}
 							trigger={
 								<Button
 									variant="primary"
@@ -331,113 +153,7 @@ function CampaignOutlook(props: CampaignOutlookProps) {
 									Donate to this campaign
 								</Button>
 							}
-						>
-							<form
-								className="flex flex-col gap-4"
-								onSubmit={(e) => {
-									e.preventDefault();
-									handleSubmit(onDonateSubmit)(e);
-								}}
-							>
-								<p className="mt-2 text-center">
-									Donate to this campaign. Every penny brings us one step closer
-									to making a difference
-								</p>
-								<div className="space-y-1 bg-white">
-									<label
-										htmlFor="donorName"
-										className="text-sm font-medium md:text-base"
-									>
-										Full name
-									</label>
-									<Input
-										{...register("donorName")}
-										autoFocus
-										type="text"
-										id="donorName"
-										required
-										placeholder="Enter your Full name"
-										className={`min-h-[45px] border text-sm font-light disabled:cursor-not-allowed disabled:bg-inherit md:text-base ${
-											errors.donorName &&
-											"ring-2 ring-abeg-error-20 placeholder:text-abeg-error-20"
-										}`}
-										disabled={user?.firstName && user.lastName ? true : false}
-									/>
-									<FormErrorMessage
-										error={errors.donorName!}
-										errorMsg={errors.donorName?.message!}
-									/>
-								</div>
-								<div className="space-y-1">
-									<label
-										htmlFor="donorEmail"
-										className="text-sm font-medium md:text-base"
-									>
-										Email address
-									</label>
-									<Input
-										{...register("donorEmail")}
-										autoFocus
-										type="text"
-										id="donorEmail"
-										required
-										placeholder="Enter your email address"
-										className={`min-h-[45px] border text-sm font-light disabled:cursor-not-allowed disabled:bg-inherit md:text-base ${
-											errors.donorEmail &&
-											"ring-2 ring-abeg-error-20 placeholder:text-abeg-error-20"
-										}`}
-										disabled={user?.email ? true : false}
-									/>
-									<FormErrorMessage
-										error={errors.donorEmail!}
-										errorMsg={errors.donorEmail?.message!}
-									/>
-								</div>
-								<div className="space-y-1">
-									<label
-										htmlFor="amount"
-										className="text-sm font-medium md:text-base"
-									>
-										Amount
-									</label>
-									<Input
-										{...register("amount")}
-										autoFocus
-										type="number"
-										id="amount"
-										required
-										placeholder="Enter any amount"
-										className={`min-h-[45px] border text-sm font-light disabled:cursor-not-allowed disabled:bg-inherit md:text-base ${
-											errors.amount &&
-											"ring-2 ring-abeg-error-20 placeholder:text-abeg-error-20"
-										}`}
-									/>
-									<FormErrorMessage
-										error={errors.amount!}
-										errorMsg={errors.amount?.message!}
-									/>
-								</div>
-								<div className="flex items-center gap-3">
-									<Checkbox
-										name="hideMyDetails"
-										id="hideMydetails"
-										className="accent-red-500"
-										checked={hideMyDetails}
-										onCheckedChange={() => setHideMyDetails(!hideMyDetails)}
-									/>
-									<label
-										htmlFor="hideMyDetails"
-										className="cursor-pointer text-sm font-medium md:text-base"
-										onClick={() => setHideMyDetails(!hideMyDetails)}
-									>
-										Donate anonymously
-									</label>
-								</div>
-								<Button className="bg-abeg-primary text-base text-white">
-									Donate
-								</Button>
-							</form>
-						</CustomDialog>
+						/>
 
 						<ShareCampaignDialog
 							campaign={campaign}
@@ -487,14 +203,26 @@ function CampaignOutlook(props: CampaignOutlookProps) {
 						>
 							Story
 						</Heading>
+
 						<div
-							className="mt-6 min-h-16 text-justify text-base md:text-lg"
+							className={cn(
+								"mt-6 min-h-16 text-justify text-base md:text-lg",
+								hideExcessStory && "line-clamp-[10]"
+							)}
 							dangerouslySetInnerHTML={{
 								__html: campaign.storyHtml,
 							}}
 						/>
 
-						<p className="mt-6 text-base md:text-xl lg:mt-12">
+						<Button
+							onClick={() => setHideExcessStory((prev) => !prev)}
+							variant="secondary"
+							className="ml-auto mt-2 shrink-0 rounded-md border-abeg-primary py-2 text-base font-medium text-abeg-primary lg:rounded-lg"
+						>
+							{hideExcessStory ? "Read more" : "Read less"}
+						</Button>
+
+						<p className="mt-8 text-base md:text-xl lg:mt-12">
 							Campaign closes on: {format(campaignDeadline, "dd-MM-yyyy")}.
 						</p>
